split setup steps into separate files
