+++ redaxo/src/core/pages/setup.step7.php
@@ -1,0 +1,31 @@
+<?php
+
+$configFile = rex_path::coreData('config.yml');
+$config = array_merge(
+    rex_file::getConfig(rex_path::core('default.config.yml')),
+    rex_file::getConfig($configFile)
+);
+$config['setup'] = false;
+
+if (rex_file::putConfig($configFile, $config)) {
+    $errmsg = '';
+    rex_file::delete(rex_path::coreCache('config.yml.cache'));
+} else {
+    $errmsg = rex_i18n::msg('setup_701');
+}
+
+$headline = rex_view::title(rex_i18n::msg('setup_700'));
+
+$content = '<h3>' . rex_i18n::msg('setup_703') . '</h3>';
+$content .= rex_i18n::rawMsg('setup_704', '<a href="' . rex_url::backendController() . '">', '</a>');
+$content .= '<p>' . rex_i18n::msg('setup_705') . '</p>';
+
+$buttons = '<a class="btn btn-setup" href="' . rex_url::backendController() . '">' . rex_i18n::msg('setup_706') . '</a>';
+
+echo $headline;
+
+$fragment = new rex_fragment();
+$fragment->setVar('heading', rex_i18n::msg('setup_702'), false);
+$fragment->setVar('body', $content, false);
+$fragment->setVar('buttons', $buttons, false);
+echo $fragment->parse('core/page/section.php');
